postgres/cdc/requirements.txt
contourpy==1.2.0
cycler==0.12.1
fonttools==4.47.2
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.3
packaging==23.2
pandas==2.1.4
pillow==10.2.0
psycopg2-binary==2.9.6
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
seaborn==0.13.1
six==1.16.0
SQLAlchemy==2.0.25
tenacity==8.2.3
typing_extensions==4.9.0
tzdata==2023.4
